Export syllabify into its own module
